+++ crates/bsc/tasks/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "reth-bsc-consensus"
+name = "reth-bsc-tasks"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -25,6 +25,8 @@ reth-engine-primitives.workspace = true
 reth-network-p2p.workspace = true
 reth-network-peers.workspace = true
 reth-beacon-consensus.workspace = true
+reth-bsc-consensus.workspace = true
+reth-evm-bsc.workspace = true
 
 # eth
 alloy-rlp.workspace = true
@@ -61,5 +63,7 @@ rand = "0.8.5"
 [features]
 bsc = [
     "reth-primitives/bsc",
-    "reth-consensus-common/bsc"
+    "reth-consensus-common/bsc",
+    "reth-bsc-consensus/bsc",
+    "reth-evm-bsc/bsc",
 ]
